feat(nuxt-styleguide-config): rename "variables" to "design-tokens"
BREAKING CHANGE
Not all sass variables are atomic and can be treated as design tokens.
Therefore the folder previosly named "variables" is renamed to "design-tokens"
and the configuration for the folder name has changed from "variablesName" to
"designTokenName".

Migration: you should rename the "variables" folder to "design-tokens"
You could also set the configuration of "designTokenName" to "variables" (not recommended)
If you allready had a custom "variablesName" configured, change the config key
to "designTokenName"