fix(nuxt-styleguide-renderer-default): namespace layout components
in order to avoid generic classnames being styled by
global styles from the project CSS

Example: Bootstrap uses a `.nav` class that would
apply its styles. With this change, the classes used
are scoped with `nsg-` to avoid these cases.

This also renames the default template to not interfere
with a projects default layout.